% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_stats.R
\name{pokemon_stats}
\alias{pokemon_stats}
\title{Determine the stats of an input pokemon}
\usage{
pokemon_stats(pokemon_data, pokemon_name)
}
\arguments{
\item{pokemon_data}{The pokemon database written by Rounak Banik}

\item{pokemon_name}{The name of the pokemon who's stats need to be analysed, starting with a Capital letter}
}
\value{
a plot containing the HP, atk, def, sp.atk, sp.def and speed of the input pokemon
}
\description{
Determine the stats of an input pokemon
}
\examples{
data<-read_csv("./pokemon.csv")
pokemon_stats(data, "Hippowdon")
}
