% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_type.R
\name{pokemon_type}
\alias{pokemon_type}
\title{Determine the type of a pokemon}
\usage{
pokemon_type(pokemon_data, ...)
}
\arguments{
\item{pokemon_data}{The pokemon database written by Rounak Banik}

\item{...}{Individual names of pokemon, starting with a Capital letter}
}
\value{
Multiple strings of text stating what types the inserted pokemon are
}
\description{
Determine the type of a pokemon
}
\examples{
data<-read_csv("./pokemon.csv")
pokemon_data(data, "Chimchar", "Drapion")
[1] "Chimchar is a fire type"
[1] "Drapion is a poison and a dark type"
}
