% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_weakness.R
\name{pokemon_weakness}
\alias{pokemon_weakness}
\title{Visualising a pokemon's weakness' and resistances}
\usage{
pokemon_weakness(pokemon_name)
}
\arguments{
\item{pokemon_name}{Name of the target pokemon}
}
\value{
A graph containing 18 pokemon types and their effectiveness on the input pokemon
}
\description{
Visualising a pokemon's weakness' and resistances
}
\examples{
pokemon_weakness("Drapion")
}
