% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon_type.R
\name{pokemon_type}
\alias{pokemon_type}
\title{Determine the type of a pokemon}
\usage{
pokemon_type(...)
}
\arguments{
\item{...}{Individual names of pokemon, starting with a Capital letter}
}
\value{
Multiple strings of text stating what types the inserted pokemon are
}
\description{
Determine the type of a pokemon
}
\examples{
pokemon_type("Chimchar", "Drapion")
}
